pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "microbiome_ml"
version = "0.1.0"
description = "Python module for standardizing ML applications in microbiome research"
authors = [
    {name = "Joshua Mitchell", email = "mitcheja@qut.edu.au"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]
dependencies = [
    "numpy>=1.21.0",
    "polars>=0.20.0",
    "requests>=2.25.0",
    "tqdm>=4.64.0",
    "pyyaml>=0.2.5",
]

[project.optional-dependencies]
geo = [
    # Optional geospatial dependencies for annotation modules
    "netcdf4>=1.6.0",
    "rasterio>=1.3.0",
    "cdsapi>=0.6.0",
    "geopandas>=0.13.0",
    "pandas>=1.5.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
    "hypothesis>=6.0.0",
]

# [project.urls]
# Homepage = "https://github.com/your-org/microbiomeML"
# Repository = "https://github.com/your-org/microbiomeML"
# Documentation = "https://your-org.github.io/microbiomeML"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 79
target-version = ["py39"]

[tool.isort]
profile = "black"
line_length = 79
multi_line_output = 3

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true
files = ["src/"]

[[tool.mypy.overrides]]
module = [
    "netcdf4.*",
    "rasterio.*",
    "cdsapi.*",
    "geopandas.*",
    "fiona.*",
    "shapely.*",
    "gdal.*",
    "pyproj.*",
    "cartopy.*",
    "regionmask.*",
    "tqdm.*",
]
ignore_missing_imports = true

[tool.flake8]
max-line-length = 79
extend-ignore = ["E203", "W503", "D", "E501"]
per-file-ignores = ["**/conftest.py:E501"]

[tool.pytest.ini_options]
testpaths = ["test/tests"]
pythonpath = ["src"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=src/microbiome_ml --cov-report=term-missing --cov-report=html"
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "geo: marks tests requiring geospatial dependencies",
]
